对代码进行优化。方便开发使用。

+++ SpringBootDev/src/main/java/org/springbootdev/core/secure/utils/SecureUtil.java
@@ -80,7 +80,7 @@ public class SecureUtil {
 			return null;
 		}
 		String clientId = Func.toStr(claims.get(SecureUtil.CLIENT_ID));
-		Integer userId = Func.toInt(claims.get(SecureUtil.USER_ID));
+		Long userId = Func.toLong(claims.get(SecureUtil.USER_ID));
 		String tenantId = Func.toStr(claims.get(SecureUtil.TENANT_ID));
 		String roleId = Func.toStr(claims.get(SecureUtil.ROLE_ID));
 		String account = Func.toStr(claims.get(SecureUtil.ACCOUNT));
@@ -103,7 +103,7 @@ public class SecureUtil {
 	 *
 	 * @return userId
 	 */
-	public static Integer getUserId() {
+	public static Long getUserId() {
 		SystemUser user = getUser();
 		return (null == user) ? -1 : user.getUserId();
 	}
@@ -114,7 +114,7 @@ public class SecureUtil {
 	 * @param request request
 	 * @return userId
 	 */
-	public static Integer getUserId(HttpServletRequest request) {
+	public static Long getUserId(HttpServletRequest request) {
 		SystemUser user = getUser(request);
 		return (null == user) ? -1 : user.getUserId();
 	}
